More robust anonymous substitution
For some reason, the version of doT on my Windows 10 box returns the anonymous function as "function anonymous(it\n/**/)" (yes, with a newline before the empty comment), so the simple string replace doesn't work and I end up with an empty JST object.  The regular expression catches everything in between the parens and so works for that case.